第一个常量指令，编译器雏形完成，编译为字节码，虚拟机执行部分明天写

+++ src/main/java/me/mathyj/exception/code/OpcodeUndefinedException.java
@@ -1,9 +1,9 @@
 package me.mathyj.exception.code;
 
 public class OpcodeUndefinedException extends RuntimeException {
-    private final byte op;
+    private final int op;
 
-    public OpcodeUndefinedException(byte op) {
+    public OpcodeUndefinedException(int op) {
         this.op = op;
     }
